+++ Yetda_iOS/Source/Storyboard/Base.lproj/Main.storyboard
@@ -72,10 +72,20 @@
                     <view key="view" contentMode="scaleToFill" id="FH2-fv-1ek">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="dtv-cY-4wp">
+                                <rect key="frame" x="184" y="694" width="46" height="30"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <state key="normal" title="Button"/>
+                            </button>
+                        </subviews>
                         <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
                         <viewLayoutGuide key="safeArea" id="CZQ-BJ-83x"/>
                     </view>
                     <navigationItem key="navigationItem" id="zFk-3v-aaa"/>
+                    <connections>
+                        <outlet property="nextButton" destination="dtv-cY-4wp" id="N2a-Hr-lOf"/>
+                    </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="0bZ-9f-ZuL" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
