+++ Yetda_iOS/Source/Views/BirthdayView/BirthdayViewController.swift
@@ -11,7 +11,8 @@
 import UIKit
 import SnapKit
 
-class BirthdayViewController: BaseViewController, UIPickerViewDelegate, UIPickerViewDataSource {
+class BirthdayViewController: BaseViewController, UIPickerViewDataSource {
+    @IBOutlet weak var nextButton: UIButton!
     
     var months: [String]! = []
     var days: [String]! = []
@@ -21,10 +22,35 @@ class BirthdayViewController: BaseViewController, UIPickerViewDelegate, UIPicker
         super.viewDidLoad()
 
         populateData()
-        setDatePicker()
+        
+//        datePicker.delegate = self
+//        datePicker.dataSource = self
         // Do any additional setup after loading the view.
     }
+    
+    // custom setup
+    override func setup() {
+        super.setup()
+    }
+    
+    // MARK: - apply layout from extension
+    override func setupUI() {
+        super.setupUI()
+        
+        setSNP()
+    }
+    
+    // populate months and days array
+    func populateData() {
+        for month in 1...12 {
+            months.append(String("\(month)월"))
+        }
 
+        for day in 1...31 {
+            days.append(String("\(day)일"))
+        }
+    }
+    
     /*
     // MARK: - Navigation
